refactor: refactor overridden Blacklight helper methods
