+++ app/helpers/layout_helper.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module LayoutHelper
+  include Blacklight::LayoutHelperBehavior
+
+  ##
+  # Classes used for sizing the main content of a Blacklight page
+  # @return [String]
+  def main_content_classes
+    "col-md-8 col-lg-9"
+  end
+
+  ##
+  # Classes used for sizing the sidebar content of a Blacklight page
+  # @return [String]
+  def sidebar_classes
+    "page-sidebar col-md-4 col-lg-3"
+  end
+end
